Some more tweaking
